Extract turn event helpers
